requirements.txt
Django==4.1.5
django-webpack-loader==1.8.0
djangorestframework==3.14.0
hubspot-api-client==7.5.0
requests==2.28.2